+++ src/main.scss
@@ -16,8 +16,10 @@
 @import 'styles/blog';
 @import 'styles/feedback';
 @import 'styles/counter';
-// @import 'styles/insights';
-@import 'styles/insights2';
+//
+@import 'styles/insights-old';
+@import 'styles/insights';
+//
 @import 'styles/interface';
 @import 'styles/missing';
 @import 'styles/resume';
