+++ src/styles/insights-old.scss
@@ -1,0 +1,254 @@
+
+$icon-facebook: ic-facebook-24px($white);
+$icon-download: ic-file-download-24px($white);
+
+$icon-check-circle: ic-check-circle-24px($white);
+$icon-check-circle-outline: ic-check-circle-outline-24px($white);
+
+// $icon-thumb-up: ic-thumb-up-24px($white);
+$icon-delete: ic-delete-forever-24px($white);
+
+.insights-old {
+  background: $dark-400;
+
+  a {
+    margin: 12px auto 0;
+    padding: 12px 18px;
+    border-radius: 3px;
+    color: $dark-400;
+    color: $white;
+    font: 600 16px/1 sans-serif;
+    background: $facebook-blue;
+    cursor: pointer;
+
+    &.-delete {
+      padding: 12px 12px 12px (12px + 24px + 6px);
+      background: $facebook-blue url($icon-delete) 12px no-repeat;
+    }
+
+    &.-disable {
+      opacity: 0.5;
+      background-color: mix($dark-400, $white, 75%);
+      pointer-events: none;
+    }
+
+    &.-facebook {
+      padding: 12px 12px 12px 48px;
+      background: $facebook-blue url($icon-facebook) 12px/30px no-repeat;
+    }
+
+    &.-import {
+      padding: 12px 12px 12px 48px;
+      background: $facebook-blue url($icon-download) 12px/30px no-repeat;
+    }
+  }
+
+  .prompt {
+    display: flex;
+    flex-flow: column;
+    width: 320px;
+    min-height: 196px;
+    margin: auto;
+    padding: 24px;
+    background: $dark-400;
+    user-select: none;
+  }
+
+  .chart {
+    background: $dark-500;
+
+    &.-overlay {
+      &::after {
+        display: block;
+        position: absolute;
+        top: 0;
+        width: 100%;
+        height: 100%;
+        background: rgba($facebook-blue, 0.25);
+        content: '';
+      }
+    }
+  }
+
+  .posts-row {
+    display: flex;
+    justify-content: space-around;
+    padding: 12px;
+
+    &:nth-child(odd) {
+      background: $dark-500;
+    }
+
+    div {
+      flex: 1;
+      text-align: right;
+    }
+  }
+}
+
+@media (min-width: 960px) {
+  .insights-old {
+    max-width: 960px;
+    margin: auto;
+    padding: 24px;
+    border-radius: 6px;
+
+    .posts-row {
+      padding: 12px 96px;
+    }
+  }
+}
+
+@media (min-width: 590px) {
+  .insights-old {
+    padding: 24px;
+
+    .prompt-wrap {
+      display: flex;
+      position: absolute;
+      z-index: 1;
+      top: 0;
+      right: 0;
+      bottom: 0;
+      left: 0;
+    }
+
+    .chart {
+      border-radius: 6px;
+      overflow: hidden;
+    }
+
+    .posts-row {
+      padding: 12px 48px;
+
+      &:nth-child(odd) {
+        border-radius: 3px;
+      }
+    }
+  }
+}
+
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+
+.insights-old {
+  .account-list {
+    padding: 18px 0 12px 0;
+  }
+
+  .account-item {
+    display: flex;
+    padding: 12px (18px + 24px) 12px 18px;
+    border-radius: 6px;
+    background: $dark-500 url($icon-check-circle-outline) right 18px center no-repeat;
+    cursor: pointer;
+
+    &.-active {
+      background: $facebook-blue url($icon-check-circle) right 18px center no-repeat;
+    }
+  }
+
+  .account-avatar {
+    // width: 72px;
+    // height: 72px;
+    // border-radius: 36px;
+    width: 48px;
+    height: 48px;
+    border-radius: 24px;
+    background-size: cover;
+  }
+
+  .account-details {
+    display: flex;
+    flex-flow: column;
+    padding: 0 0 0 18px;
+
+    h1 {
+      margin: auto 0 0;
+      font: 600 18px/20px sans-serif;
+    }
+
+    div {
+      margin: 0 0 auto;
+      font: 400 13px/18px sans-serif;
+    }
+  }
+
+  //
+
+  .prompt-avatar {
+    width: 128px;
+    height: 128px;
+    margin: 0 auto 24px;
+    border-radius: 64px;
+    background-size: cover;
+  }
+
+  .prompt-body {
+    h1 {
+      font: 600 24px/1 sans-serif;
+      text-align: center;
+    }
+
+    p {
+      margin: 12px 0 6px;
+      font: 400 14px/24px sans-serif;
+    }
+  }
+
+  //
+  //
+  //
+
+  //
+  //
+  //
+
+  .avatar {
+    width: 72px;
+    height: 72px;
+    margin: 0 18px 0 0;
+    border: 6px solid $dark-400;
+    border-radius: 36px;
+    background-size: cover;
+
+    &.-instagram {
+      margin: 0 18px 0 (0 - 18px - 18px - 6px);
+    }
+  }
+
+  .insights-head {
+    display: flex;
+    // min-height: 90px;
+    // padding: 0 6px 18px;
+    padding: 0 0 18px;
+    font: 600 16px/64px sans-serif;
+
+    a {
+      margin: auto 0;
+
+      &.-refresh {
+        margin: auto 0 auto auto;
+      }
+
+      &.-logout {
+        margin: auto 0 auto 12px;
+      }
+    }
+  }
+
+  //
+
+  .posts {
+    min-height: 120px;
+    padding: 24px 0 0;
+  }
+}
